requirements.txt
google_oauth2_tool==0.0.3
google-cloud-translate==3.12.0
